Health bars, effectively done
Now NPCs and Heros both have health bars, provided in the new Lib.py.
Rendering for those also interpolates color as damage increases, so that
nearly dead = red, fully alive = green, between = between. Fixed a
couple minor bugs along the way. Had gaps between the color and black
parts of the bar, now truncating width to int to prevent gaps. Had a bug
with handling on NPC health values where the check didn't match the init
names. At any rate, all is now working.
